ostree-ext: Serialize xattrs into tar stream as well
We really want this for https://github.com/coreos/rpm-ostree/pull/5222
to be able to rebuild images from their container-synthesized rootfs.

Really, the only xattr we don't want to emit in to the tar stream
is security.selinux for now.

Eventually we should try to switch to putting that into the tar
stream too, but it needs more validation.

Signed-off-by: Colin Walters <walters@verbum.org>